Change the hash calculation method.
Previously, the nodehash and parenthash were computed
using the node representation. Unfortunately, this method of computation
was sensitive to minor changes (unsignificant spaces etc).

This works similarly to the jumble query mechanism in pg_stat_statements, by
keeping only the relevant info on each node, for a subset of known nodes.

For other nodes, we fall back to the node representation.