Order quals and constants by their position in text.
In the plan tree, quals can appear in an order different from the one used in
the text query. This proves problematic when trying to re-inject constant
values in the "jumbled" query. This small change order them by their token
position in the original text.

Thanks to github user shribe for reporting it.