Manage differences between database encodings.
Since pg_qualstats() can be called from anywhere, and will likely end up
in a powa database, difference in encoding must be taken into account.

For constants, every one of them is stored internally as utf8, decoded from
the current database encoding. At call time, the utf8 string is encoded in
the current database encoding, which might or might not be compatible with the
original encoding (ie, unrepresentable code points). If the whole cluster is in
UTF-8, this is basically a no-op.