Allow pgqs to work locally without being in shared_preload_libraries.
In certain cases, some people would prefer to run pg_qualstats without having
it loaded in all backends and/or having to restart postgres, but would still
want to have it run on production data to see the exact behavior of their
queries.  The use case is quite narrow, as one would need to be able to run all
the wanted queries in a row in the same backend to benefit from that feature,
but there's some demand for that and the impact in the code to be able to
support that is not that high.

Per request from Magnus Hagander.