Add PGDLLEXPORT on exposed functions.
The PG_FUNCTION_INFO_V1() marco also takes care of the extra annotation for the
generated info function but not the exposed function itself.

Similarly, the _PG_init() and _PG_fini() functions needs to be exported
(although the _PG_fini() is currently not called by postgres).

This should fix the runtime errors on Windows reported by github user suprimex.