Fix a faulty strncpy() call.
As this strncpy() was done against a hardcoded set of values known to be
smaller than the target pointer this couldn't lead to any buffer overflow, but
this was still wrong.

As the set of values will never change, simply add an assert to check the
source size as a safeguard, and use a plain strcpy().

Per report from github user RekGRpth.