Fix extension relocability.
This extension isn't fully relocatable.  It internally needs to reference the
target schema explicitly, so a custom schema can only be set as installation
time.

Mark it as not relocatable in the control file and properly handle the schema
qualification in the install script.

Regression tests are updated to validate that the extension work when installed
in a custom and not visible schema.