Add minimal CI.
This is a basic github action setup.  It's currently executed only on the
master branch and PRs against it.  It's run against all currrent postgres major
versions on ubuntu 22.04.  Note that the list of major versions and the OS
version are hardcoded and have to be maintained.

Adapted from powa-archivist.