Fix crash when a special var points to another special var
When a var has a special varno (ie a negative varno), we need to resolve the
actual underlying Var, as we would otherwise end up reading random memory when
trying to access the associated RangeTblEntry.  The previous code was correctly
resolving such Var, but didn't consider the possibility that the resolved Var
could itself have a special varno which needs to be resolved again.

Thanks to Kenny Chen for the report.